+++ docs/release_notes/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 UNRELEASED
 
 - added support for :class:`pandas.Timedelta`, :class:`datetime.timedelta` and :class:`numpy.timedelta64`
+- added :meth:`staircase.Stairs.from_values` (#GH88)
 
 Please list new changes above this comment
 
@@ -62,7 +63,7 @@ Contributors: @amagee (Andrew Magee)
 
 - bugfix for missing *lower_how* and *upper_how* parameters in :meth:`staircase.Stairs.resample`
 - renamed Stairs._values to avoid error when performing Series.groupby.sum in pandas>=1.2
- 
+
 
 **v1.6.2 2020-01-13**